+++ demisto_sdk/commands/common/content/objects_factory.py
@@ -9,7 +9,8 @@ from demisto_sdk.commands.common.content.objects.pack_objects import (
     GenericField, GenericModule, GenericType, IncidentField, IncidentType,
     IndicatorField, IndicatorType, Integration, Layout, LayoutsContainer,
     OldClassifier, OldIndicatorType, PackIgnore, PackMetaData, Playbook,
-    Readme, ReleaseNote, Report, Script, SecretIgnore, Widget)
+    Readme, ReleaseNote, ReleaseNoteConfig, Report, Script, SecretIgnore,
+    Widget)
 from demisto_sdk.commands.common.content.objects.root_objects import \
     Documentation
 from demisto_sdk.commands.common.tools import find_type
@@ -40,6 +41,7 @@ TYPE_CONVERSION_BY_FileType = {
     FileType.DESCRIPTION: Readme,
     FileType.README: Readme,
     FileType.RELEASE_NOTES: ReleaseNote,
+    FileType.RELEASE_NOTES_CONFIG: ReleaseNoteConfig,
     FileType.DOC_IMAGE: DocFile,
     FileType.JAVASCRIPT_FILE: '',
     FileType.POWERSHELL_FILE: '',