+++ demisto_sdk/commands/generate_test_playbook/test_playbook_generator.py
@@ -327,12 +327,18 @@ class PlaybookTestsGenerator:
         else:
             input_folder = Path(input)
 
-            if 'Packs' in (p.name for p in input_folder.parents):
-                """
-                if input yml is under standard Packs/<Pack>/Integrations/<Integration> path,
-                save the test-playbook under   Packs/<Pack>/TestPlaybooks
-                """
-                folder = (input_folder.parent.parent / 'TestPlaybooks')
+            """
+            if input yml is under standard Packs/<Pack>/<...>/<Integration> path,
+            save the test-playbook under   Packs/<Pack>/TestPlaybooks
+            """
+            pack_path = None
+            for p in input_folder.parents:
+                if p.parent.name == "Packs":
+                    pack_path = p
+                    break
+
+            if pack_path:
+                folder = (pack_path / 'TestPlaybooks')
                 folder.mkdir(exist_ok=True, parents=True)
             else:
                 """ otherwise, save the generated test-playbook in the folder from which SDK is called."""