+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * Enhanced ***generate-docs*** with default `additionalinformation` (description) for common parameters.
 * Added to **validate** command a validation that a content item's id and name will not end with spaces.
 * The **format** command will now remove trailing whitespaces from content items' id and name fields.
+* Fixed an issue where the **generate-test-playbook** command would not place the playbook in the proper folder.
+* Added to **validate** command a validation that packs with `Iron Bank` uses the latest docker from Iron Bank.
+* Added to the **create artifacts** command support for release notes configuration file.
+* Added validation to **validate** for release notes config file.
 
 # 1.4.8
 * Fixed an issue where yml files with `!reference` failed to load properly.